+++ skorch/tests/test_helper.py
@@ -254,6 +254,207 @@ class TestSliceDict:
         assert sldict0 != sldict1
 
 
+class TestSliceDataset:
+    @pytest.fixture(scope='class')
+    def data(self):
+        X, y = make_classification(100, 20, n_informative=10, random_state=0)
+        return X.astype(np.float32), y
+
+    @pytest.fixture
+    def X(self, data):
+        return data[0]
+
+    @pytest.fixture
+    def y(self, data):
+        return data[1]
+
+    @pytest.fixture
+    def custom_ds(self, data):
+        """Return a custom dataset instance"""
+        from skorch.dataset import Dataset
+        class MyDataset(Dataset):
+            """Simple pytorch dataset that returns 2 values"""
+            def __len__(self):
+                return len(self.X)
+
+            def __getitem__(self, i):
+                Xi = self.X[i]
+                yi = self.y[i]
+                return self.transform(Xi, yi)
+
+        return MyDataset(*data)
+
+    @pytest.fixture(scope='session')
+    def slds_cls(self):
+        from skorch.helper import SliceDataset
+        return SliceDataset
+
+    @pytest.fixture
+    def slds(self, slds_cls, custom_ds):
+        return slds_cls(custom_ds)
+
+    @pytest.fixture
+    def slds_y(self, slds_cls, custom_ds):
+        return slds_cls(custom_ds, idx=1)
+
+    def test_len_and_shape(self, slds, y):
+        assert len(slds) == len(y)
+        assert slds.shape == (len(y),)
+
+    @pytest.mark.parametrize('sl', [
+        slice(0, 1),
+        slice(1, 2),
+        slice(0, 2),
+        slice(0, None),
+        slice(-1, None),
+        slice(None, None, -1),
+        [0],
+        [55],
+        [-3],
+        [0, 10, 3, -8, 3],
+        np.ones(100, dtype=np.bool),
+        # boolean mask array of length 100
+        np.array([0, 0, 1, 0] * 25, dtype=np.bool),
+    ])
+    def test_len_and_shape_sliced(self, slds, y, sl):
+        assert len(slds[sl]) == len(y[sl])
+        assert slds[sl].shape == (len(y[sl]),)
+
+    @pytest.mark.parametrize('n', [0, 1])
+    def test_slice_non_int_is_slicedataset(self, slds_cls, custom_ds, n):
+        slds = slds_cls(custom_ds, idx=n)
+        sl = np.arange(7, 55, 3)
+        sliced = slds[sl]
+        assert isinstance(sliced, slds_cls)
+        assert np.allclose(sliced.indices_, sl)
+
+    @pytest.mark.parametrize('n', [0, 1])
+    @pytest.mark.parametrize('sl', [0, 55, -3])
+    def test_slice(self, slds_cls, custom_ds, X, y, sl, n):
+        data = y if n else X
+        slds = slds_cls(custom_ds, idx=n)
+        sliced = slds[sl]
+        x = data[sl]
+        assert np.allclose(sliced, x)
+
+    @pytest.mark.parametrize('n', [0, 1])
+    @pytest.mark.parametrize('sl0, sl1', [
+        ([55], 0),
+        (slice(0, 1), 0),
+        (slice(-1, None), 0),
+        ([55], -1),
+        ([0, 10, 3, -8, 3], 1),
+        (np.ones(100, dtype=np.bool), 5),
+        # boolean mask array of length 100
+        (np.array([0, 0, 1, 0] * 25, dtype=np.bool), 6),
+    ])
+    def test_slice_twice(self, slds_cls, custom_ds, X, y, sl0, sl1, n):
+        data = X if n == 0 else y
+        slds = slds_cls(custom_ds, idx=n)
+        sliced = slds[sl0][sl1]
+        x = data[sl0][sl1]
+        assert np.allclose(sliced, x)
+
+    @pytest.mark.parametrize('n', [0, 1])
+    @pytest.mark.parametrize('sl0, sl1, sl2', [
+        (slice(0, 50), slice(10, 20), 5),
+        ([0, 10, 3, -8, 3], [1, 2, 3], 2),
+        (np.ones(100, dtype=np.bool), np.arange(10, 40), 29),
+    ])
+    def test_slice_three_times(self, slds_cls, custom_ds, X, y, sl0, sl1, sl2, n):
+        data = y if n else X
+        slds = slds_cls(custom_ds, idx=n)
+        sliced = slds[sl0][sl1][sl2]
+        x = data[sl0][sl1][sl2]
+        assert np.allclose(sliced, x)
+
+    def test_explicitly_pass_indices_at_init(self, slds_cls, custom_ds, X):
+        # test passing indices directy to __init__
+        slds = slds_cls(custom_ds, indices=np.arange(10))
+        sliced0 = slds[5:]
+        assert np.allclose(sliced0, X[5:10])
+
+        sliced1 = sliced0[2]
+        assert np.allclose(sliced1, X[7])
+
+    def test_access_element_out_of_bounds(self, slds_cls, custom_ds):
+        slds = slds_cls(custom_ds, idx=2)
+        with pytest.raises(IndexError) as exc:
+            # pylint: disable=pointless-statement
+            slds[0]
+
+        msg = ("SliceDataset is trying to access element 2 but there are only "
+               "2 elements.")
+        assert exc.value.args[0] == msg
+
+    def test_fit_with_slds_works(self, slds, y, classifier_module):
+        from skorch import NeuralNetClassifier
+        net = NeuralNetClassifier(classifier_module)
+        net.fit(slds, y)  # does not raise
+
+    def test_fit_with_slds_without_valid_works(self, slds, y, classifier_module):
+        from skorch import NeuralNetClassifier
+        net = NeuralNetClassifier(classifier_module, train_split=False)
+        net.fit(slds, y)  # does not raise
+
+    def test_grid_search_with_slds_works(
+            self, slds, y, classifier_module):
+        from sklearn.model_selection import GridSearchCV
+        from skorch import NeuralNetClassifier
+
+        net = NeuralNetClassifier(
+            classifier_module,
+            train_split=False,
+            verbose=0,
+        )
+        params = {
+            'lr': [0.01, 0.02],
+            'max_epochs': [10, 20],
+        }
+        gs = GridSearchCV(net, params, refit=False, cv=3, scoring='accuracy', iid=True)
+        gs.fit(slds, y)  # does not raise
+
+    def test_grid_search_with_slds_and_internal_split_works(
+            self, slds, y, classifier_module):
+        from sklearn.model_selection import GridSearchCV
+        from skorch import NeuralNetClassifier
+
+        net = NeuralNetClassifier(classifier_module)
+        params = {
+            'lr': [0.01, 0.02],
+            'max_epochs': [10, 20],
+        }
+        gs = GridSearchCV(net, params, refit=True, cv=3, scoring='accuracy', iid=True)
+        gs.fit(slds, y)  # does not raise
+
+    def test_grid_search_with_slds_X_and_slds_y(
+            self, slds, slds_y, classifier_module):
+        from sklearn.model_selection import GridSearchCV
+        from skorch import NeuralNetClassifier
+
+        net = NeuralNetClassifier(
+            classifier_module,
+            train_split=False,
+            verbose=0,
+        )
+        params = {
+            'lr': [0.01, 0.02],
+            'max_epochs': [10, 20],
+        }
+        gs = GridSearchCV(net, params, refit=False, cv=3, scoring='accuracy', iid=True)
+        gs.fit(slds, slds_y)  # does not raise
+
+    def test_index_with_2d_array_raises(self, slds):
+        i = np.arange(4).reshape(2, 2)
+        with pytest.raises(IndexError) as exc:
+            # pylint: disable=pointless-statement
+            slds[i]
+
+        msg = ("SliceDataset only supports slicing with 1 "
+               "dimensional arrays, got 2 dimensions instead.")
+        assert exc.value.args[0] == msg
+
+
 class TestPredefinedSplit():
 
     @pytest.fixture